[feature][refactor] global style injection using emotion global component; imported project relevant font families using a css import statement at aforementioned global style file; improved folder structure to better organize the global mui theme
