+++ libs/frontend-theme/.storybook/tsconfig.json
@@ -15,5 +15,5 @@
     "../**/*.spec.tsx",
     "../**/*.spec.jsx"
   ],
-  "include": ["../src/**/*", "*.js"]
+  "include": ["../src/**/*", "*.js", "../index.d.ts"]
 }
