+++ libs/frontend-theme/src/lib/GlobalStyles/Global.tsx
@@ -1,0 +1,13 @@
+import { Global } from '@emotion/react';
+
+const GlobalStyles = () => {
+  return (
+    <Global
+      styles={`
+  @import url('https://fonts.googleapis.com/css2?family=Nunito+Sans:ital,wght@0,200;0,300;0,400;0,600;0,700;0,800;0,900;1,200;1,300;1,600;1,700;1,800;1,900&display=swap');
+  `}
+    />
+  );
+};
+
+export default GlobalStyles;
