+++ libs/frontend-theme/src/lib/theme/breakpoints.ts
@@ -1,0 +1,12 @@
+import { ThemeOptions } from '@mui/material';
+
+const breakpoints: ThemeOptions['breakpoints'] = {
+  values: {
+    xs: 300,
+    sm: 640,
+    md: 1024,
+    lg: 1170,
+    xl: 1400,
+  },
+};
+export default breakpoints;
