+++ libs/frontend-theme/src/lib/theme/palette.ts
@@ -1,0 +1,12 @@
+import { ThemeOptions } from '@mui/material';
+
+const palette: ThemeOptions['palette'] = {
+  primary: {
+    main: '#FF793F',
+  },
+  secondary: {
+    main: '#283047',
+  },
+};
+
+export default palette;
