+++ libs/frontend-theme/src/lib/theme/theme.ts
@@ -1,12 +1,6 @@
 import { createTheme, responsiveFontSizes } from '@mui/material';
-import { TypographyStyleOptions } from '@mui/material/styles/createTypography';
-
-const titleStyle = (size: string): TypographyStyleOptions => ({
-  fontSize: size,
-  fontFamily: 'Nunito Sans, sans-serif',
-  fontStyle: 'normal',
-  color: '#323232',
-});
+import breakpoints from './breakpoints';
+import typography from './typography';
 
 export const palette = createTheme({
   palette: {
@@ -38,27 +32,8 @@ export const theme = createTheme({
     },
   },
 
-  breakpoints: {
-    values: {
-      xs: 300,
-      sm: 640,
-      md: 1024,
-      lg: 1170,
-      xl: 1400,
-    },
-  },
-  typography: {
-    htmlFontSize: 10,
-    h1: titleStyle('33px'),
-    h2: titleStyle('26px'),
-    h3: titleStyle('22px'),
-    h4: titleStyle('18px'),
-    h5: titleStyle('16px'),
-    h6: titleStyle('12px'),
-    body1: {
-      color: '#757575',
-    },
-  },
+  breakpoints: breakpoints,
+  typography: typography,
 
   palette: {
     primary: {
